Adapt 'inflected' interface.
The 'inflected' library can only convert from snake case to the other
cases. It also has a bit of an oddball interface that departs from
ActiveSupport. The `camelize` method takes a boolean to see if it
should capitalize the first letter.

This adds an layer to present the expected behaviors to the rest of
mirage.